Also make the tracing regs structure test more autoconfy.
Don't use the _ONCE check for the ptrace.h header since on Linux we can
stop once we found linux/ptrace.h (two compiler calls less than the usual,
in the best case scenario).

And make sure to die during the configure step rather than in the build
when no structure to trace regs is found.